# marketstack-openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Marketstack OpenAPI",
    "description": "Inofficial marketstack OpenAPI spec",
    "version": "1.0"
  },
  "paths": {
    "/eod": {
      "get": {
        "tags": [
          "eod"
        ],
        "summary": "Query",
        "operationId": "eod",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.EodPrice__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/eod/latest": {
      "get": {
        "tags": [
          "eod"
        ],
        "summary": "Latest",
        "operationId": "eod_latest",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.EodPrice__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/eod/{date}": {
      "get": {
        "tags": [
          "eod"
        ],
        "summary": "Date",
        "operationId": "eod_date",
        "parameters": [
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": true,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_List_models.EodPrice__"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/tickers/{symbol}/eod": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol Eod",
        "operationId": "ticker_symbol_eod",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "exchange",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_TickerEod_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickers/{symbol}/eod/latest": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol Eod Latest",
        "operationId": "ticker_symbol_eod_latest",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EodPrice"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickers/{symbol}/eod/{date}": {
      "get": {
        "tags": [
          "tickers"
        ],
        "summary": "Symbol Eod Date",
        "operationId": "ticker_symbol_eod_date",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            },
            "example": "AAPL",
            "name": "symbol",
            "in": "path"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": true,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EodPrice"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/exchanges/{mic}/eod": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Mic Eod",
        "operationId": "exchange_mic_eod",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "mic",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Date/time sort order",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Sort"
                }
              ]
            },
            "example": "DESC",
            "name": "sort",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_from",
            "in": "query"
          },
          {
            "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z",
            "required": false,
            "schema": {
              "title": "Timestamp",
              "type": "string",
              "description": "Date in the formats %Y-%m-%d, %Y-%m-%d %H:%M:%S or ISO-8601 %Y-%m-%dT%H:%M:%S+%Z"
            },
            "example": "2022-11-05T00:00:00+0000",
            "name": "date_to",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 1000.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "example": 10,
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer"
            },
            "example": 0,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_ExchangeEod_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exchanges/{mic}/eod/latest": {
      "get": {
        "tags": [
          "exchanges"
        ],
        "summary": "Mic Eod Latest",
        "operationId": "exchange_mic_eod_latest",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Exchange MIC",
              "type": "string"
            },
            "example": "XNAS",
            "name": "mic",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "API access key",
              "type": "string"
            },
            "name": "access_key",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Comma-separated symbols list",
              "type": "string"
            },
            "example": "AAPL,AMZN",
            "name": "symbols",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_ExchangeEod_"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EodPrice": {
        "title": "EodPrice",
        "required": [
          "date",
          "symbol",
          "open",
          "low",
          "high",
          "exchange",
          "split_factor",
          "dividend",
          "adj_open",
          "adj_close",
          "adj_high",
          "adj_low",
          "adj_volume"
        ],
        "type": "object",
        "properties": {
          "date": {
            "title": "Date",
            "type": "string",
            "format": "date-time"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "volume": {
            "title": "Volume",
            "type": "number"
          },
          "open": {
            "title": "Open",
            "type": "number"
          },
          "close": {
            "title": "Close",
            "type": "number"
          },
          "low": {
            "title": "Low",
            "type": "number"
          },
          "high": {
            "title": "High",
            "type": "number"
          },
          "exchange": {
            "title": "Exchange",
            "type": "string"
          },
          "last": {
            "title": "Last",
            "type": "number"
          },
          "split_factor": {
            "title": "Split Factor",
            "type": "number"
          },
          "dividend": {
            "title": "Dividend",
            "type": "number"
          },
          "adj_open": {
            "title": "Adj Open",
            "type": "number"
          },
          "adj_close": {
            "title": "Adj Close",
            "type": "number"
          },
          "adj_high": {
            "title": "Adj High",
            "type": "number"
          },
          "adj_low": {
            "title": "Adj Low",
            "type": "number"
          },
          "adj_volume": {
            "title": "Adj Volume",
            "type": "number"
          }
        }
      },
      "Error": {
        "title": "Error",
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "context": {
            "title": "Context",
            "type": "object"
          }
        }
      },
...
      "ErrorResponse": {
        "title": "ErrorResponse",
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "PagedResponse_ExchangeEod_": {
        "title": "PagedResponse[ExchangeEod]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "$ref": "#/components/schemas/ExchangeEod"
          }
        }
      },
...
      "PagedResponse_List_models.EodPrice__": {
        "title": "PagedResponse[List[models.EodPrice]]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "title": "Data",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EodPrice"
            }
          }
        }
      },
...
      "PagedResponse_TickerEod_": {
        "title": "PagedResponse[TickerEod]",
        "required": [
          "pagination",
          "data"
        ],
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "$ref": "#/components/schemas/TickerEod"
          }
        }
      },
...
      "Pagination": {
        "title": "Pagination",
        "required": [
          "limit",
          "offset",
          "count",
          "total"
        ],
        "type": "object",
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        }
      },
      "Sort": {
        "title": "Sort",
        "enum": [
          "ASC",
          "DESC"
        ],
        "description": "An enumeration."
      },
...
    }
  }
}